Refactoring Step 2: Updating standalone_make utility code
